practice adds

+++ Week4-PracticePE4B_temp/PE12_temp.cpp
@@ -7,7 +7,7 @@ using namespace std;
 int main() {
 
     double money, change;
-    const double quarter = 0.25;
+    const float quarter = 0.25;
 
     cout << "Enter the amount of money: $";
     cin >> money;
